wishlist#14 added multi game picker
