+++ client/static/app/modules/games/main.js
@@ -10,16 +10,15 @@ define(function(require) {
       GamesDetailsController = require('./controllers/game-details.controller'),
       GamesDetailsTpl = require('text!./templates/game-details.tpl.html'),
 
-      GamePickerDirective = require('./directives/game-picker.directive'),
-
       GameService = require('./services/game.service'),
 
+      MultiGamePickerDirective = require('./directives/multi-game-picker.directive'),
       TypeAheadDirective = require('./directives/type-ahead.directive'),
       module = angular.module('GamesModule', []);
 
-  module.directive('dgGamePicker', GamePickerDirective);
   module.factory('dgGameService', GameService);
   module.directive('dgTypeAhead', TypeAheadDirective);
+  module.directive('dgMultiGamePicker', MultiGamePickerDirective);
 
   GamesModuleConfig.$inject = ['$routeProvider'];
   module.config(GamesModuleConfig);
