--- client/static/app/modules/games/templates/game-picker.tpl.html
@@ -1,19 +1,0 @@
-<div class="choose-game">
-
-  <ui-select ng-model="dgGamePickerIns.selectedGames.selected" theme="select2"
-             on-select="dgGamePickerIns.onCitySelect($item, $model)">
-
-    <ui-select-match placeholder="Выберите игру">{{$select.selected.name}}</ui-select-match>
-
-    <ui-select-choices repeat="game in dgGamePickerIns.games | filter: $select.search"
-                       refresh="dgGamePickerIns.getGames()" refresh-delay="0">
-
-      <div ng-bind-html="dgGamePickerIns.selectedGames.name | highlight: $select.search"></div>
-
-      {{game.name}}
-
-    </ui-select-choices>
-
-  </ui-select>
-
-</div>
