+++ client/static/app/modules/wishlist/templates/wishlist-page.tpl.html
@@ -22,7 +22,7 @@
           <h3>Игры</h3>
         </div>
         <div class="ibox-content">
-          <dg-game-picker></dg-game-picker>
+          <dg-multi-game-picker></dg-multi-game-picker>
         </div>
       </div>
     </div>
